requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Werkzeug==3.0.2
pandas==2.2.2
matplotlib==3.9.0